Fix sprite y value issues
If a sprite started off the screen a complicated series of checks needed to be done to see if the sprite intersected the current scanline.

The new method decides which sprites to include instead of which to exclude, making the if statement more clear.

This fixes sprites getting chopped off at the top of the screen.